Added Author class. All TG classes moved to 1 file
Added an Author class. Updated all typegoose classes to appropriately
reference or use a generic array of X class (it was incorrect before).
Also consolidated all classes into a Classes.ts file. This is
necessary to avoid a dependency cycle, where the classses import each
other so they can be used with mongoose.populate(). And since the
classes will never actually be called outside of Models.ts where it is
used to build schemas, this is a good compromise.